WEB-1981 - statistics
